@@ -17,9 +17,10 @@ package android.example.com.squawker.following;
 
 import android.example.com.squawker.R;
 import android.os.Bundle;
-import android.support.v4.app.NavUtils;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.NavUtils;
 import android.view.MenuItem;
 
 /**
