@@ -18,12 +18,13 @@ package android.example.com.squawker;
 
 import android.database.Cursor;
 import android.example.com.squawker.provider.SquawkContract;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
